Separate Habit from CategoryForm ToDo: Show name of the habit in Habit component
