+++ src/CategoryForm.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import styled from '@emotion/styled';
 
-import skillBook from './images/skillBook.jpeg';
+import Habit from './Habit';
 
 import GlobalExample from './GlobalExample';
 
@@ -27,13 +27,6 @@ const Container = styled.div({
   fontSize: '1.6em',
   textAlign: 'center',
 });
-
-const HabitImage = styled.div({
-  margin: '4% 0 auto',
-  height: '100',
-  width: '300',
-  textAlign: 'center',
-})
 
 const Instruction = styled.div({
   margin: '8% 2% 1% 2%',
@@ -66,12 +59,7 @@ export default function CategoryForm({ categories, onChange, onClick }) {
       <Container>
         <p>축하합니다! 습관을 형성했습니다!</p>
       </Container>
-      <HabitImage>
-        <img
-          src={skillBook}
-          width="225"
-        />
-      </HabitImage>
+      <Habit/>
       <Instruction>
         <p>습관을 다양한 세 종류로 나누어보세요</p>
       </Instruction>
