+++ src/Habit.test.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import { render } from '@testing-library/react';
+
+import Habit from './Habit';
+
+test('Habit', () => {
+  render(
+      <Habit />
+    );
+});
